requirements.txt
asttokens==2.4.1
blinker==1.7.0
click==8.1.7
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
dnspython==2.4.2
executing==2.0.1
Flask==3.0.0
Flask-PyMongo==2.3.0
ipykernel==6.28.0
ipython==8.19.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jsonify==0.5
jupyter_client==8.6.0
jupyter_core==5.6.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
packaging==23.2
parso==0.8.3
pexpect==4.9.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pyzmq==25.1.2
six==1.16.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.1
wcwidth==0.2.12
Werkzeug==3.0.1
